Fixed a bug where not all references would be inputted into the database
To be fair: this bug isn't fixed. It was intentional.
Documents that do not have DOIs labelled as: {10.XYZ} (WITH THE FIRST THREE CHARACTERS BEING EXACTLY "10.") are not added to the list of references. Because of this, the arraylist of SLRs could read

Article 1
Article 3
Article 4
Article 5

If article 2 had an improperly formatted doi. This is intentional for the time being but is important to note. There is now a println that displays if an article has an improperly formatted doi, though it is still not added to the list of references.